requirements.txt
pandas
numpy
nltk
scikit-learn
streamlit
Pillow
